#3 #11 #6
-Main class: outsourced business logic to models; shortened code;
leveraged the currentPlayer property and nextplayer method of the game
class to iterate through the players.

-Field: created FieldState-enum and added a getState-method to the Field
class in order to improve the readability of a fields' state in the main
class. The nested if-else-statement is now buried within the model.

-Game: added nextPlayer method; the gameover method now also sets the
winner of a game to prevent duplicate execution of code; added a
getEnemiesOfCurrentPlayer-method to outsource business logic from the
main class

-Player: added selectShip-method that gives immediate feedback to the
player if the chosen ship can be selected; it also helps outsourcing
business logic from the main class

-started adding javadoc comments where useful